Update generated files to fix bug https://bitbucket.org/geckofx/geckofx-5.0/issue/7

+++ GeckofxUnitTests/FocusManagerTests.cs
@@ -74,12 +74,12 @@ namespace GeckofxUnitTests
 			throw new NotImplementedException();
 		}
 
-		public double GetTextZoomAttribute()
+		public float GetTextZoomAttribute()
 		{
 			throw new NotImplementedException();
 		}
 
-		public void SetTextZoomAttribute(double aTextZoom)
+		public void SetTextZoomAttribute(float aTextZoom)
 		{
 			throw new NotImplementedException();
 		}
